// src/api/model/stockTransactionsCreateBodyType.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Inventory API
 * Inventory & Stock Management REST API — Repository + Service pattern, domain events, HMAC device webhooks, audit trail.
 * OpenAPI spec version: 1.0.0
 */

export type StockTransactionsCreateBodyType = typeof StockTransactionsCreateBodyType[keyof typeof StockTransactionsCreateBodyType];


export const StockTransactionsCreateBodyType = {
  in: 'in',
  out: 'out',
  adjustment: 'adjustment',
} as const;